+++ PROJECT-Rogue/Assets/Scripts/Weapons/Projectiles/ProjectileSniperRifle.cs
@@ -8,7 +8,7 @@ public class ProjectileSniperRifle : Projectile
     {
         var obj = Object.Instantiate(Shooting.instance.bulletPrefabs.Find(x => x.tag.Contains("|ProjSniper|")), whoAttacks.firePoint.position, whoAttacks.firePoint.rotation);
         Vector2 ownerVelocity = CalculateOwnersVelocity((whoAttacks.GetComponent<Rigidbody2D>().velocity / 6), 2f, 3f);
-        obj.GetComponent<Bullet>().SetParameters(whoAttacks, damage * whoAttacks.Damage, bulletSpeed, bulletSize, ownerVelocity, (whoAttacks.Range * rangeModifier) / 10);
+        obj.GetComponent<Bullet>().SetParameters(whoAttacks, (int)damage * (int)whoAttacks.Damage, bulletSpeed, bulletSize, ownerVelocity, (whoAttacks.Range * rangeModifier) / 10);
     }
     //public ProjectileSniperRifle(float attackSpeed = 0.4f, int damage = 50, float bulletSpeed = 30, float bulletSize = 1) : base(attackSpeed, damage, bulletSpeed, bulletSize) { }
